refactor: split converter.py utilities into content_utils.py
Extract 13 module-level functions (~350 lines) to content_utils.py.
MarkdownConverter class remains in converter.py. Re-exports kept in
converter.py for backward compatibility.
