+++ src/scraper/base/__init__.py
@@ -1,4 +1,4 @@
-from src.scraper.base.converter import (
+from src.scraper.base.content_utils import (
     calc_pages,
     clean_markdown,
     clean_norm_soup,
@@ -9,6 +9,7 @@ from src.scraper.base.converter import (
     valid_markdown,
     wrap_html,
 )
+from src.scraper.base.converter import MarkdownConverter  # noqa: F401
 from src.scraper.base.scraper import (
     DEFAULT_LLM_PROMPT,
     BaseScraper,
@@ -22,6 +23,7 @@ from src.utils.concurrency import run_in_thread
 __all__ = [
     "BaseScraper",
     "DEFAULT_LLM_PROMPT",
+    "MarkdownConverter",
     "run_in_thread",
     "is_pdf",
     "detect_extension",
